chore(ui): refactor `WdsTextInputLayout` - AB-60
