+++ src/ui/src/builder/settings/BuilderTemplateInputInput.vue
@@ -1,17 +1,14 @@
 <template>
-	<div
+	<WdsTextInputLayout
+		:left-icon
+		:right-icon
+		:invalid
+		:variant
 		class="BuilderTemplateInputInput"
-		:class="{
-			'BuilderTemplateInputInput--ghost': variant === 'ghost',
-			'BuilderTemplateInputInput--multiline': multiline,
-		}"
-		:aria-invalid="invalid"
-		:style="{
-			gridTemplateColumns: gridTemplateColumns,
-		}"
+		:class="{ 'BuilderTemplateInputInput--multiline': multiline }"
+		@right-icon-click="$emit('rightIconClick')"
 		@click="focus"
 	>
-		<i v-if="leftIcon" class="material-symbols-outlined">{{ leftIcon }}</i>
 		<SharedContentEditable
 			ref="input"
 			class="BuilderTemplateInputInput__input"
@@ -23,21 +20,14 @@
 				:background-colors="backgroundTagColors"
 			/>
 		</SharedContentEditable>
-		<button
-			v-if="rightIcon"
-			class="BuilderTemplateInputInput__rightIcon"
-			type="button"
-			@click="$emit('rightIconClick')"
-		>
-			<i class="material-symbols-outlined">{{ rightIcon }}</i>
-		</button>
-	</div>
+	</WdsTextInputLayout>
 </template>
 
 <script setup lang="ts">
 import { computed, inject, onMounted, PropType, useTemplateRef } from "vue";
 import StateWithPills from "../stateDropdown/StateWithPills.vue";
 import SharedContentEditable from "@/components/shared/SharedContentEditable.vue";
+import WdsTextInputLayout from "@/wds/WdsTextInputLayout.vue";
 import { useDynamicUserState } from "../useDynamicUserState";
 import injectionKeys from "@/injectionKeys";
 import { extractObjectPaths } from "@/utils/object";
@@ -97,16 +87,6 @@ onMounted(() => {
 	if (props.autofocus) focus();
 });
 
-const gridTemplateColumns = computed(() =>
-	[
-		props.leftIcon ? "auto" : undefined,
-		"minmax(0, 1fr)",
-		props.rightIcon ? "auto" : undefined,
-	]
-		.filter(Boolean)
-		.join(" "),
-);
-
 async function onChange(value: string) {
 	model.value = value;
 	emit("input", {
@@ -141,64 +121,6 @@ function focus() {
 </script>
 
 <style scoped>
-.BuilderTemplateInputInput {
-	width: 100%;
-	margin: 0;
-	border: 1px solid var(--separatorColor);
-	border-radius: 8px;
-	padding: 8.5px 12px 8.5px 12px;
-	font-size: 14px;
-	outline: none;
-	color: var(--primaryTextColor);
-	background: transparent;
-	transition: box-shadow ease-in-out 0.2s;
-
-	cursor: pointer;
-	display: grid;
-	align-items: center;
-	gap: 8px;
-}
-
-.BuilderTemplateInputInput--multiline {
-	resize: vertical;
-}
-
-.BuilderTemplateInputInput:focus-within {
-	border: 1px solid var(--softenedAccentColor);
-	box-shadow: 0px 0px 0px 3px rgba(81, 31, 255, 0.05);
-}
-
-.BuilderTemplateInputInput--ghost {
-	border-color: transparent;
-	background-color: transparent;
-	transition:
-		box-shadow,
-		background-color,
-		border-color ease-in-out 0.2s;
-}
-.BuilderTemplateInputInput--ghost:disabled {
-	cursor: not-allowed;
-}
-.BuilderTemplateInputInput--ghost:not(:disabled):hover {
-	background-color: transparent;
-	border-color: var(--wdsColorBlue3);
-}
-.BuilderTemplateInputInput--ghost:not(:disabled):focus,
-.BuilderTemplateInputInput--ghost:not(:disabled):focus-within {
-	background-color: transparent;
-}
-
-.BuilderTemplateInputInput i {
-	color: var(--wdsColorGray5);
-}
-
-.BuilderTemplateInputInput__rightIcon {
-	border: none;
-	background-color: transparent;
-	display: flex;
-	align-items: center;
-	cursor: pointer;
-}
 .BuilderTemplateInputInput__input {
 	background: transparent;
 	font-size: 14px;
@@ -207,15 +129,10 @@ function focus() {
 	min-height: 21px;
 	height: 100%;
 }
-.BuilderTemplateInputInput--multiline .BuilderTemplateInputInput__input {
-	min-height: 64px;
-}
 .BuilderTemplateInputInput__input:focus {
-	border: none;
-	box-shadow: none;
 	outline: none;
 }
-.BuilderTemplateInputInput[aria-invalid="true"] {
-	border-color: var(--wdsColorOrange5);
+.BuilderTemplateInputInput--multiline .BuilderTemplateInputInput__input {
+	min-height: 64px;
 }
 </style>
